Preliminary version of different LnF for nested tabs.  May need some UI love.
Also, maybe want to control this by a nested=true parameter (similar to tabStrip=true parameter); currently it triggers automatically for any nested tabs.

There's a quirk in the underline on IE due to the img.dijitTabButtonSpacer; that node should be moved to before the text or after the icon, or preferably removed, if possible.

Also, if you have a ContentPane than includes a TabContainer via href, should set padding to 0 to make it look good.

Fixes #7628. [[16029]]