Defer focusing on a static Menu/MenuBar until user clicks it, or tabs into it.   Once user clicks on a Menu/MenuBar, it focuses on it, and then (as before) any mouse movement or keyboard movement (via arrow keys) will change focus.   Also (as before), focus is shifted to a context menu immediately when it's displayed.
Separated out CSS classes for indicating that a MenuItem is mouse hovered (dijitMenuItemHover), and to indicate which MenuItem is selected/active (dijitMenuItemSelected).   Currently they look exactly the same, although that could be changed (including removing the hover effect altogether).

"Selected/active" is in the sense of the selected tab, and is controlled by the mouse or keyboard.  Implementation-wise, it means that either the MenuItem has focus, or focus is on a submenu of that MenuItem.

Also added dijitMenuPassive/dijitMenuActive class to the Menu/MenuBar domNode so that CSS rules for hover can be customized based on whether or not the menu has focus.   Once the menu gets focus the dijitMenuHover effect is disabled in favor of the dijitMenuSelected effect, so that the dijitMenuHover effect won't linger on "File" if user moved the mouse over "File" but then used the keyboard arrows to move to the "Edit" MenuBarItem.  (This is a setting in tundra/nihilo/soria and can be changed if desired.)

Open issue with re-focus on click-execute of a leaf menu item... after executing a menu choice focus ends up nowhere (or maybe on the page itself).   Probably should be going to the menubar?

Refs #4634 !strict [[16233]]