Fixes #8561: Visible IFRAMEs when using Dijits on IE6
I removed the code to pre-create a background iframe on IE6 on page load. This code was added in [10099] (refs #4057) to prevent a 2s delay on IE6 showing the first popup, but I'm not seeing that problem anymore, and the code is questionable anyway since it slows down page load time. [[16540]]