Fixes #8710: make onHide() work for ContentPane.
Make onHide() and onShow() callbacks available for all children of StackContainer (and it's subclasses, TabContainer and AccordionContainer), and for all ContentPane  subclasses that are shown/hidden: Dialog, TooltipDialog, and TitlePane.

Note that unfortunately there are competing standards: onShow/onHide are called in the cases mentioned above, but onOpen/onClose are called by the dijit.popup code, for example when a Menu is popped up, or a Calendar is dropped down from a DateTextBox.  onOpen/onClose are also called for TreeNodes.  I'd like to standardize that naming for 2.0.

!strict [[17400]]