Fix problem (regression?) where search list wasn't being filtered by IME input.
On IE and safari, IME's (for inputting kanji etc.) produce keycode = 229 for keystrokes while in the IME input mode, and then no special indication when the input is finished and the IME has inserted some characters into the <input> box.

Firefox, OTOH, has a compositionend event to indicate when the IME has inserted some characters into the input, and we handle that by producing a faux call to _onKeyPress() with no keycode defined at all.

Fixes #9437 !strict. [[18282]]