Various TabContainer href-loading and children resizing fixes, centered around ContentPane's role as a layout widget:
  - Fix doLayout=false TabContainer (broken by recent checkins for #5528, #9794)
  - Fix problem where resizing the TabContainer would refresh a pane with refreshOnShow=true.  It's only supposed to refresh when the pane is hidden then shown again.  (fixes #8969 !strict)

Differentiated between onShow(), which indicates that the ContentPane is being shown, and resize(), which, when the ContentPane is in a layout widget hierarchy, occurs both when the ContentPane is shown but also when it changes size.

It's complicated because a ContentPane being made visible is indicated in different ways depending on the situation:

  - For a TitlePane, opening the TitlePane just generates an _onShow() notification
  - For a TabContainer --> BorderContainer --> ContentPane configuration, the ContentPane just gets a resize() call when first made visible
  - For the typical TabContainer --> ContentPane, it gets both an _onShow() and a resize() call

Refs #5528, #9794 !strict

Also added a bunch of tests including one for a TitlePane with a single layout child (refs #8197). [[20099]]