Code to deregister iframes/main window to avoid memory leaks on IE.   Also, about mousedown listener, switched to using attachEvent()/attachListener() rather than dojo.connect() because the latter leaves a dojo._ieDispatcher() handler on dojo.doc (even after the dojo.disconnect() call), which causes a leak warning.
This change solves the leak, although could probably come up with a better design for the handle returned from registerWin()/registerIframe().   But that's an implementation detail.  I was also conflicted about whether unregisterIframe() should take a handle or a reference to the iframe; ended up passing in a handle, like dojo.disconnect().

Refs #10208 !strict. [[20662]]