Switch !TreeNode to extend _CssStateMixin.   Doesn't change too much since we still need the event handlers in !TreeNode so it can notify Tree about those events.
Note that the hover and focus events are just for the rowNode, not for the whole !TreeNode (ie: they don't includes the children of the !TreeNode).

Changed dijitTreeNodeHover to dijitTreeRowHover and dijitTreeNodeSelected to dijitTreeRowSelected since those classes are applied to the .dijitTreeRow DOMNode (they only affect the !TreeNode itself, not the !TreeNode's children).

Refs #10613 !strict. [[21148]]