Refactoring Dialog to handle when multiple dialogs are displayed and a non-top-level dialog is hidden or destroyed.  The salient change is that rather than a Dialog saving the previous focus, there's a "level manager" that remembers each active layer (starting with the base page itself) and the focus in that layer.
Also fixes race conditions on concurrent dialog show/hide.

Previously Dialog._onKey was connected to the document body.   I couldn't find a case when that was necessary (as opposed to connecting to Dialog.domNode), so I changed it to attach to Dialog.domNode, so Dialog wouldn't have to check "am I the top dialog on the stack?" in that method.

Fixes #2238, #9944, #10705 !strict. [[22662]]