Tooltip positioning fixes:
   - Before positioning tooltip, limit tooltip's width to available width between the around node and the edge of the viewport.   Reducing a tooltip's width (for tooltips with flowing text) will increase the tooltip's height, which affects vertical placement.

   - Position tooltip connector arrow to line up with the around node, even when tooltip (due to a large tooltip size/small viewport size) extends both above and below the around node.

   - Fix problem with tooltip placement when clicking a !TextBox causes it to scroll into view simultaneously with displaying a tooltip.

Note: To determine the "natural" width of a toolip, had to remove left:50% setting for offscreen tooltips, because offscreen tooltips still get limited in width to the edge of the viewport.   That CSS was apparently to work around a bug in an old browser (but no longer happens in the browsers we support).

Patch from Katie Vance (IBM, CCLA), thanks!

Fixes #4997, #4282 !strict.

Known issues: regression test failing by a few pixels on Mac FF (but working on Windows FF). [[22969]]