Support assigning multiple panes to the same !BorderContainer region, so !BorderContainer can be used in place of !SplitContainer or !LayoutContainer without requiring nested !BorderContainer widgets.
Panes can specify a "layoutPriority" attribute that controls which panes are closer to the edges of the !BorderContainer whenever two panes have the same region, or top and left / bottom and right panes are competing.

Also includes a bug fix for display of dragged bottom/right splitter when it's moved beyond its maximum allowed value.

Fixes #11030 !strict. [[23443]]