Support for _mapXXXAttr attributes that specifying a mapping from a widget's XXX attribute to a DOMNode attribute/style/innerHTML etc.
Each _mapXXXAttr attribute/value is equivalent to a single key/value pair from attributeMap.  The _mapXXXAttr API supercedes attributeMap, which is deprecated and will be removed in 2.0.

For backwards-compatibility, the widget attributes with attributeMap entries are processed first, followed by the attributes with _mapXXXAttr objects and _setXXXAttr() methods.   The _mapXXXAttr objects / _setXXXAttr() methods are called according to the order the attributes are declared in the widget.

Refs #12278 !strict. [[23941]]